+++ backend/prisma/migrations/20260909032235_add_enabled_features/migration.sql
@@ -1,0 +1,2 @@
+-- AlterTable
+ALTER TABLE "ChurchSettings" ADD COLUMN     "enabledFeatures" JSONB NOT NULL DEFAULT '{}';